Update remote when only using periodic pull
At this point in time the only (builtin) way of pushing local changes is when "push after commit" is active. But this can lead to errors when there are remote changes (since you always have to pull before you can push).

This change adds a pull to the periodic task to push changes. 

Alternatively the "push after commit" could be extended to always pull in changes before pushing.